refactor(inputs.sqlserver): moved azureToken struct in a new file
